Athenas_Lite/athenas_web/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
reportlab==4.2.2
aiofiles==24.1.0
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
requests==2.32.3
python-dotenv==1.0.0
google-generativeai==0.8.3
pandas==2.2.3
soundfile==0.12.1
mutagen==1.47.0
